Change WIFI_FEATURE_XXX type from int to long
This change is to avoid issues in assigning signed integer
feature (WIFI_FEATURE_DPP) to long featureSet. The
assignement was resulting in large, signed integer.

Bug: 143363031
Test: atest com.android.wifi.server
Test: Wifi ON/OFF, Connect and data transfer
Test: Added logs and verified feature set saved in iface.
(cherry picked from commit 18b0feceb73b1465635b0153e4cf88eb51ce4614)

+++ tests/wifitests/src/com/android/server/wifi/ClientModeImplTest.java
@@ -1789,10 +1789,10 @@ public class ClientModeImplTest {
     /** Verifies that syncGetSupportedFeatures() masks out capabilities based on system flags. */
     @Test
     public void syncGetSupportedFeatures() {
-        final int featureAware = WifiManager.WIFI_FEATURE_AWARE;
-        final int featureInfra = WifiManager.WIFI_FEATURE_INFRA;
-        final int featureD2dRtt = WifiManager.WIFI_FEATURE_D2D_RTT;
-        final int featureD2apRtt = WifiManager.WIFI_FEATURE_D2AP_RTT;
+        final long featureAware = WifiManager.WIFI_FEATURE_AWARE;
+        final long featureInfra = WifiManager.WIFI_FEATURE_INFRA;
+        final long featureD2dRtt = WifiManager.WIFI_FEATURE_D2D_RTT;
+        final long featureD2apRtt = WifiManager.WIFI_FEATURE_D2AP_RTT;
         final long featureLongBits = 0x1100000000L;
 
         assertEquals(0, testGetSupportedFeaturesCase(0, false));
